Create a helper class to make queries eaiser
