+++ .pg_format
@@ -2,3 +2,4 @@ tabs=1
 function-case=1 #lowercase
 keyword-case=1
 type-case=1
+no-space-function=1
